Stop using bitpay's CreateInvoice for non bitpay API usage
